+++ pkg/http/api/client/errors.go
@@ -2,40 +2,23 @@ package client
 
 import (
 	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
 	"net/http"
-	"strings"
+
+	"github.com/b3scale/b3scale/pkg/http/api"
 )
-
-// APIError will contain the decoded json body
-// when the response status was not OK or Accepted
-type APIError map[string]interface{}
-
-// Error implements the error interface
-func (err APIError) Error() string {
-	errs := []string{}
-	for k, v := range err {
-		errs = append(errs, fmt.Sprintf("%s: %v", k, v))
-	}
-	return strings.Join(errs, "; ")
-}
 
 // ErrRequestFailed creates an APIError from the
 // HTTP response.
-func ErrRequestFailed(res *http.Response) APIError {
+func ErrRequestFailed(res *http.Response) api.ServerError {
 	defer res.Body.Close()
 	body, err := io.ReadAll(res.Body)
 	if err != nil {
-		return APIError{"error": err.Error()}
+		return api.ServerError{"error": err.Error()}
 	}
-	apiErr := APIError{}
+	apiErr := api.ServerError{}
 	if err := json.Unmarshal(body, &apiErr); err != nil {
-		return APIError{"error": body}
+		return api.ServerError{"error": body}
 	}
 	return apiErr
 }
-
-// ErrNotFound is the response when a response is a 404
-var ErrNotFound = errors.New("the resource could not be found (404)")
